- Reorganização dos pactoes - Criação do container com Redis, Mysql, RabbitMQ - Ajustes no Swagger
