+++ src/main/java/com/santanna/serviceorder/domain/dto/OrderRequestDto.java
@@ -1,4 +1,4 @@
-package com.santanna.serviceorder.dto;
+package com.santanna.serviceorder.domain.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.DecimalMin;
