+++ src/main/java/com/santanna/serviceorder/infrastructure/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.santanna.serviceorder.config;
+package com.santanna.serviceorder.infrastructure.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
