+++ src/main/java/com/santanna/serviceorder/infrastructure/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.santanna.serviceorder.config;
+package com.santanna.serviceorder.infrastructure.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
